# Conflicts:
#	Session.xcodeproj/project.pbxproj
#	Session/Conversations/ConversationVC+Interaction.swift
#	Session/Conversations/ConversationViewModel.swift
#	Session/Media Viewing & Editing/GIFs/GifPickerViewController.swift
#	Session/Media Viewing & Editing/GIFs/GiphyAPI.swift
#	Session/Meta/Translations/de.lproj/Localizable.strings
#	Session/Meta/Translations/en.lproj/Localizable.strings
#	Session/Meta/Translations/es.lproj/Localizable.strings
#	Session/Meta/Translations/fa.lproj/Localizable.strings
#	Session/Meta/Translations/fi.lproj/Localizable.strings
#	Session/Meta/Translations/fr.lproj/Localizable.strings
#	Session/Meta/Translations/hi.lproj/Localizable.strings
#	Session/Meta/Translations/hr.lproj/Localizable.strings
#	Session/Meta/Translations/id-ID.lproj/Localizable.strings
#	Session/Meta/Translations/it.lproj/Localizable.strings
#	Session/Meta/Translations/ja.lproj/Localizable.strings
#	Session/Meta/Translations/nl.lproj/Localizable.strings
#	Session/Meta/Translations/pl.lproj/Localizable.strings
#	Session/Meta/Translations/pt_BR.lproj/Localizable.strings
#	Session/Meta/Translations/ru.lproj/Localizable.strings
#	Session/Meta/Translations/si.lproj/Localizable.strings
#	Session/Meta/Translations/sk.lproj/Localizable.strings
#	Session/Meta/Translations/sv.lproj/Localizable.strings
#	Session/Meta/Translations/th.lproj/Localizable.strings
#	Session/Meta/Translations/vi-VN.lproj/Localizable.strings
#	Session/Meta/Translations/zh-Hant.lproj/Localizable.strings
#	Session/Meta/Translations/zh_CN.lproj/Localizable.strings
#	Session/Notifications/AppNotifications.swift
#	Session/Notifications/SyncPushTokensJob.swift
#	Session/Notifications/UserNotificationsAdaptee.swift
#	SessionMessagingKit/Configuration.swift
#	SessionMessagingKit/Database/Models/Interaction.swift
#	SessionMessagingKit/Database/Models/SessionThread.swift
#	SessionMessagingKit/Jobs/Types/MessageReceiveJob.swift
#	SessionMessagingKit/Jobs/Types/MessageSendJob.swift
#	SessionMessagingKit/Jobs/Types/SendReadReceiptsJob.swift
#	SessionMessagingKit/Messages/Message.swift
#	SessionMessagingKit/Sending & Receiving/Message Handling/MessageReceiver+ReadReceipts.swift
#	SessionMessagingKit/Sending & Receiving/MessageSender+Convenience.swift
#	SessionMessagingKit/Sending & Receiving/MessageSender.swift
#	SessionMessagingKit/Shared Models/MentionInfo.swift

+++ Session/Conversations/ConversationVC+Interaction.swift
@@ -7,6 +7,7 @@ import Photos
 import PhotosUI
 import Sodium
 import GRDB
+import SessionUIKit
 import SessionMessagingKit
 import SessionUtilitiesKit
 import SignalUtilitiesKit
@@ -200,6 +201,30 @@ extension ConversationVC:
     // MARK: - ExpandingAttachmentsButtonDelegate
 
     func handleGIFButtonTapped() {
+        guard Storage.shared[.isGiphyEnabled] else {
+            let modal: ConfirmationModal = ConfirmationModal(
+                info: ConfirmationModal.Info(
+                    title: "GIPHY_PERMISSION_TITLE".localized(),
+                    explanation: "GIPHY_PERMISSION_MESSAGE".localized(),
+                    confirmTitle: "continue_2".localized()
+                ) { [weak self] _ in
+                    Storage.shared.writeAsync(
+                        updates: { db in
+                            db[.isGiphyEnabled] = true
+                        },
+                        completion: { _, _ in
+                            DispatchQueue.main.async {
+                                self?.handleGIFButtonTapped()
+                            }
+                        }
+                    )
+                }
+            )
+            
+            present(modal, animated: true, completion: nil)
+            return
+        }
+        
         let gifVC = GifPickerViewController()
         gifVC.delegate = self
         
@@ -434,10 +459,17 @@ extension ConversationVC:
                     .filter(id: threadId)
                     .updateAll(db, SessionThread.Columns.shouldBeVisible.set(to: true))
                 
+                let authorId: String = {
+                    if let blindedId = self?.viewModel.threadData.currentUserBlindedPublicKey {
+                        return blindedId
+                    }
+                    return self?.viewModel.threadData.currentUserPublicKey ?? getUserHexEncodedPublicKey(db)
+                }()
+                
                 // Create the interaction
                 let interaction: Interaction = try Interaction(
                     threadId: threadId,
-                    authorId: getUserHexEncodedPublicKey(db),
+                    authorId: authorId,
                     variant: .standardOutgoing,
                     body: text,
                     timestampMs: sentTimestampMs,
@@ -789,6 +821,8 @@ extension ConversationVC:
             let actions: [ContextMenuVC.Action] = ContextMenuVC.actions(
                 for: cellViewModel,
                 recentEmojis: (self.viewModel.threadData.recentReactionEmoji ?? []).compactMap { EmojiWithSkinTones(rawValue: $0) },
+                currentUserPublicKey: self.viewModel.threadData.currentUserPublicKey,
+                currentUserBlindedPublicKey: self.viewModel.threadData.currentUserBlindedPublicKey,
                 currentUserIsOpenGroupModerator: OpenGroupManager.isUserModeratorOrAdmin(
                     self.viewModel.threadData.currentUserPublicKey,
                     for: self.viewModel.threadData.openGroupRoomToken,
@@ -839,7 +873,7 @@ extension ConversationVC:
     }
 
     func handleItemTapped(_ cellViewModel: MessageViewModel, gestureRecognizer: UITapGestureRecognizer) {
-        guard cellViewModel.variant != .standardOutgoing || cellViewModel.state != .failed else {
+        guard cellViewModel.variant != .standardOutgoing || (cellViewModel.state != .failed && cellViewModel.state != .failedToSync) else {
             // Show the failed message sheet
             showFailedMessageSheet(for: cellViewModel)
             return
@@ -1474,30 +1508,34 @@ extension ConversationVC:
     // MARK: --action handling
     
     func showFailedMessageSheet(for cellViewModel: MessageViewModel) {
-        let sheet = UIAlertController(title: cellViewModel.mostRecentFailureText, message: nil, preferredStyle: .actionSheet)
-        sheet.addAction(UIAlertAction(title: "Cancel", style: .cancel, handler: nil))
-        sheet.addAction(UIAlertAction(title: "Delete", style: .destructive, handler: { _ in
-            Storage.shared.writeAsync { db in
-                try Interaction
-                    .filter(id: cellViewModel.id)
-                    .deleteAll(db)
-            }
-        }))
-        sheet.addAction(UIAlertAction(title: "Resend", style: .default, handler: { _ in
-            Storage.shared.writeAsync { [weak self] db in
-                guard
-                    let threadId: String = self?.viewModel.threadData.threadId,
-                    let interaction: Interaction = try? Interaction.fetchOne(db, id: cellViewModel.id),
-                    let thread: SessionThread = try SessionThread.fetchOne(db, id: threadId)
-                else { return }
-                
-                try MessageSender.send(
-                    db,
-                    interaction: interaction,
-                    in: thread
-                )
-            }
-        }))
+        let sheet = UIAlertController(
+            title: (cellViewModel.state == .failedToSync ?
+                "MESSAGE_DELIVERY_FAILED_SYNC_TITLE".localized() :
+                "MESSAGE_DELIVERY_FAILED_TITLE".localized()
+            ),
+            message: cellViewModel.mostRecentFailureText,
+            preferredStyle: .actionSheet
+        )
+        sheet.addAction(UIAlertAction(title: "TXT_CANCEL_TITLE".localized(), style: .cancel, handler: nil))
+        
+        if cellViewModel.state != .failedToSync {
+            sheet.addAction(UIAlertAction(title: "TXT_DELETE_TITLE".localized(), style: .destructive, handler: { _ in
+                Storage.shared.writeAsync { db in
+                    try Interaction
+                        .filter(id: cellViewModel.id)
+                        .deleteAll(db)
+                }
+            }))
+        }
+        
+        sheet.addAction(UIAlertAction(
+            title: (cellViewModel.state == .failedToSync ?
+                "context_menu_resync".localized() :
+                "context_menu_resend".localized()
+            ),
+            style: .default,
+            handler: { [weak self] _ in self?.retry(cellViewModel) }
+        ))
         
         // HACK: Extracting this info from the error string is pretty dodgy
         let prefix: String = "HTTP request failed at destination (Service node "
@@ -1581,6 +1619,23 @@ extension ConversationVC:
     }
     
     // MARK: - ContextMenuActionDelegate
+    
+    func retry(_ cellViewModel: MessageViewModel) {
+        Storage.shared.writeAsync { [weak self] db in
+            guard
+                let threadId: String = self?.viewModel.threadData.threadId,
+                let interaction: Interaction = try? Interaction.fetchOne(db, id: cellViewModel.id),
+                let thread: SessionThread = try SessionThread.fetchOne(db, id: threadId)
+            else { return }
+            
+            try MessageSender.send(
+                db,
+                interaction: interaction,
+                in: thread,
+                isSyncMessage: (cellViewModel.state == .failedToSync)
+            )
+        }
+    }
 
     func reply(_ cellViewModel: MessageViewModel) {
         let maybeQuoteDraft: QuotedReplyModel? = QuotedReplyModel.quotedReplyForSending(
@@ -1855,10 +1910,16 @@ extension ConversationVC:
                 })
                 
                 actionSheet.addAction(UIAlertAction(
-                    title: (cellViewModel.threadVariant == .legacyGroup || cellViewModel.threadVariant == .group ?
-                        "delete_message_for_everyone".localized() :
-                        String(format: "delete_message_for_me_and_recipient".localized(), threadName)
-                    ),
+                    title: {
+                        switch cellViewModel.threadVariant {
+                            case .legacyGroup, .group: return "delete_message_for_everyone".localized()
+                            default:
+                                return (cellViewModel.threadId == userPublicKey ?
+                                    "delete_message_for_me_and_my_devices".localized() :
+                                    String(format: "delete_message_for_me_and_recipient".localized(), threadName)
+                                )
+                        }
+                    }(),
                     style: .destructive
                 ) { [weak self] _ in
                     deleteRemotely(
